+++ tablesalt/lib/tablesalt/thread_accessor/management.rb
@@ -3,26 +3,27 @@
 module Tablesalt
   module ThreadAccessor
     module Management
-      def store
-        Thread.current[Tablesalt::ThreadAccessor::STORE_THREAD_KEY] ||= ThreadStore.new
+      def store(namespace = nil)
+        stores = Thread.current[Tablesalt::ThreadAccessor::STORE_THREAD_KEY] ||= {}
+        stores[namespace] ||= ThreadStore.new
       end
 
       # Cleans up ThreadAccessor state after given block executes
       #
       # @param :logger [Logger] Optional; A logger instance that implements the method :warn to send warning messages to
       # @yield Required; Yields no variables to the given block
-      def clean_thread_context(logger: nil)
-        if store.present?
+      def clean_thread_context(logger: nil, namespace: nil)
+        if store(namespace).present?
           if logger.nil?
-            puts "WARNING: ThreadAccessor variables set outside ThreadAccessor context: #{store.keys.join(", ")}"
+            puts "WARNING: ThreadAccessor variables set outside ThreadAccessor context: #{store(namespace).keys.join(", ")}"
           else
-            logger.warn("ThreadAccessor variables set outside ThreadAccessor context: #{store.keys.join(", ")}")
+            logger.warn("ThreadAccessor variables set outside ThreadAccessor context: #{store(namespace).keys.join(", ")}")
           end
         end
 
         yield
       ensure
-        store.clear
+        store(namespace).clear
       end
     end
   end