+++ tablesalt/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming <!-- Add unreleased change notes here: -->
 
 - Isolated ThreadAccessor store, Rack middleware ([#462](https://github.com/Freshly/spicerack/pull/462))
+- Added namespaced thread stores for gem usage ([#463](https://github.com/Freshly/spicerack/pull/463))
 
 ## v0.26.0.4